floatに変更

+++ src/IR.cpp
@@ -11,11 +11,11 @@ static const int NOT_FOUND_THRESHOLD = 900;
 // =====内部状態（外から触れない）=====
 static int irValue[11];
 static bool ballFound = false;
-static double ballAngle = 0.0;
+static float ballAngle = 0.0;
 
 // =====内部関数=====
-static double deg_radian(int index){
-    static const double Deg[IR_COUNT] = {
+static float deg_radian(int index){
+    static const float Deg[IR_COUNT] = {
         0.0, 45.0, 90.0, 135.0,
         180.0, 225.0, 270.0, 315.0
     };
@@ -54,18 +54,18 @@ void IR_update()
     }
 
     // =====ベクトル合成=====
-    double vx = 0.0, vy = 0.0;
+    float vx = 0.0, vy = 0.0;
 
     for (int pin = IR_FIRST_PIN; pin <= IR_LAST_PIN; pin++)
     {
-        double w = 1.0 / (irValue[pin] + 1);
-        double ang = deg_radian(pin - IR_FIRST_PIN);
+        float w = 1.0 / (irValue[pin] + 1);
+        float ang = deg_radian(pin - IR_FIRST_PIN);
         vx += w * cos(ang);
         vy += w * sin(ang);
     }
 
     //角度計算
-    double theta = atan2(vy,vx) * M_PI / 180.0;
+    float theta = atan2(vy,vx) * M_PI / 180.0;
     if (theta < 0.0) theta += 360.0;
 
     ballAngle = theta;
@@ -80,7 +80,7 @@ bool IR_ballFound()
     return ballFound;
 }
 
-double IR_getAngle()
+float IR_getAngle()
 {
     return ballAngle;
 }
