+++ src/IR.h
@@ -3,4 +3,4 @@
 void IR_init();
 void IR_update();
 bool IR_ballFound();
-double IR_getAngle();
+float IR_getAngle();
